Resolve all warnings
Most warnings were disabled by surrounding all
third party includes with a pragma

Others were manually fixed. PVS Studio was also
ran on the project, which found a few errors.

Updated to later versions of assimp & glm

Removed unused ReflectionProbe class &
PhysicsDebuggerSettings.cpp

+++ FlexEngine/include/Graphics/GL/GLRenderer.hpp
@@ -136,7 +136,7 @@ namespace flex
 			void DrawWorldSpaceSprites(const GameContext& gameContext);
 			void DrawText(const GameContext& gameContext);
 
-			bool LoadFont(const GameContext& gameContext, BitmapFont** font, const std::string& filePath, i32 size);
+			bool LoadFont(const GameContext& gameContext, BitmapFont** font, const std::string& filePath, i16 size);
 
 			void UpdateTextBuffer();
